# Improve attach_to_template() function